requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
backoff==2.2.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clickhouse-driver==0.2.9
ConfigArgParse==1.7
Deprecated==1.2.18
fastapi==0.109.1
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
idna==3.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.33.0
MarkupSafe==3.0.2
msgpack==1.1.0
opentelemetry-api==1.22.0
opentelemetry-distro==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-instrumentation-jinja2==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-sqlite3==0.43b0
opentelemetry-instrumentation-system-metrics==0.43b0
opentelemetry-instrumentation-tortoiseorm==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
protobuf==4.25.6
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==0.19.0
pytz==2025.1
pyzmq==26.2.1
requests==2.32.3
setuptools==75.8.2
sniffio==1.3.1
starlette==0.35.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.15.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
